Add abstract class
