+++ mod-erm-usage-counter41/src/main/java/org/olf/erm/usage/counter41/csv/CSVMapper.java
@@ -1,8 +1,10 @@
 package org.olf.erm.usage.counter41.csv;
 
+import java.util.Objects;
 import java.util.stream.Stream;
 import org.niso.schemas.counter.Report;
-import org.olf.erm.usage.counter41.csv.mapper.JournalReport1Mapper;
+import org.olf.erm.usage.counter41.csv.mapper.AbstractCounterReport;
+import org.olf.erm.usage.counter41.csv.mapper.JR1;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -11,15 +13,28 @@ public class CSVMapper {
   private static final Logger LOG = LoggerFactory.getLogger(CSVMapper.class);
   private static final String[] JR1 = new String[] {"JR1", "Journal Report 1"};
 
+  private static AbstractCounterReport getType(Report report) {
+    Objects.requireNonNull(report.getVersion());
+    Objects.requireNonNull(report.getTitle());
+
+    if (report.getVersion().equals("4")) {
+      String title = report.getTitle();
+      if (Stream.of(JR1).anyMatch(title::contains)) {
+        return new JR1(report);
+      }
+    }
+    return null;
+  }
+
   public static String toCSV(Report report) {
-    if (Stream.of(JR1).anyMatch(s -> report.getTitle().contains(s))
-        && report.getVersion().equals("4")) {
-      return new JournalReport1Mapper(report).toCsv();
+    AbstractCounterReport type = getType(report);
+    if (type != null) {
+      return type.toCSV();
     } else {
       LOG.error(
-          String.format(
-              "No mapping found for report title '%s' and version '%s'",
-              report.getTitle(), report.getVersion()));
+          "No mapping found for report title '{}' and version '{}'",
+          report.getTitle(),
+          report.getVersion());
       return null;
     }
   }
